db: split db_register
