+++ backend/matcha/services/users.py
@@ -12,8 +12,9 @@ from matcha.app_utils import (
     flaskprint,
 )
 
+from matcha.db.register import db_register
+
 from matcha.db.db import (
-    db_register,
     db_set_user_email,
     db_get_user_email,
     db_get_interests,
